Updated component packages to add branch-alias
Now that we have per-component repos, we need the ability to install
dev-stability packages. Currently, this can only be done if you specify
dev-master (which is auto-enabled via packagist), but not via dev-develop.
Additionally, you cannot specify version strings like "2.1-dev" or "2.2-dev", as
those aliases do not exist.

This patch adds dev-master and dev-develop aliases for each component.